Move creation of motion model widgets into new module
